requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.3.4
    # via aiohttp
aiohttp==3.10.1
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==24.1.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via -r requirements.in
discord-py==2.4.0
    # via -r requirements.in
feedparser==6.0.11
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.7
    # via yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
python-dotenv==1.0.1
    # via -r requirements.in
sgmllib3k==1.0.0
    # via feedparser
soupsieve==2.5
    # via beautifulsoup4
yarl==1.9.4
    # via aiohttp
